South America + Mexico leaders, Amazonas flag.
Didn't add everyone, for CHL, BRA and ARG it is just ones with a portrait without DLC. ARG seems to have an extra commie guy but he was in Spain in 1936. For Mexico i added one DLC locked character, but that's it. I'm just trying to fill the capitals at this point.

Vargas' democratic role doesn't work for some reason, will invastiage that.