Use OOO for the target of each basgame tag alias without crashing.
OOO now spawns with a cosmetic tag in place of Kerguelen, which is removed after the game starts. The crash only happens when OOO is removed at the start of the game so this allows me to keep the tag aliases without any basegame code intervening with playable states.

Also other error.log cleanups.